Fixed compilation error and dependency
