+++ src/runtime_src/core/edge/drm/zocl/zocl_ioctl.c
@@ -15,19 +15,9 @@
 
 #include "sched_exec.h"
 #include "zocl_xclbin.h"
-#include "zocl_generic_cu.h"
 #include "zocl_error.h"
 
 extern int kds_mode;
-
-int zocl_graph_alloc_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-        struct sched_client_ctx *client);
-int zocl_graph_free_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-        struct sched_client_ctx *client);
-int zocl_aie_alloc_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-        struct sched_client_ctx *client);
-int zocl_aie_free_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-        struct sched_client_ctx *client);
 
 /*
  * read_axlf and ctx should be protected by zdev_xclbin_lock exclusively.
@@ -68,7 +58,7 @@ zocl_ctx_ioctl(struct drm_device *ddev, void *data, struct drm_file *filp)
 	int ret = 0;
 
 	if (kds_mode == 1) {
-		/* Do not acquire zdev_xclbin_lock like zocl_xclbin_ctx().
+		/* Do not acquire zdev_xclbin_lock like sched_xclbin_ctx().
 		 * New KDS would lock bitstream when open the fist context.
 		 * The lock bitstream would exclude read_axlf_ioctl().
 		 */
